Added timestamp support and decreased frequency of presence updates
