+++ package.json
@@ -9,7 +9,6 @@
     "discord-rich-presence": "^0.0.6",
     "fancy-log": "^1.3.2",
     "jsdom": "^11.6.2",
-    "nodemon": "^1.17.2",
     "snekfetch": "^4.0.0-beta.0"
   },
   "devDependencies": {
